scripts/requirements.txt
python-mlb-statsapi>=1.5.0
requests>=2.31.0
